chore: rate limit bot interactions
