+++ pokerapp/pokerbotview.py
@@ -27,10 +27,25 @@ from pokerapp.entities import (
     PlayerState,
 )
 
+
+class RateLimitedSender:
+    """Serializes send/edit/delete operations to avoid rate limits."""
+
+    def __init__(self, delay: float = 0.5):
+        self._delay = delay
+        self._lock = asyncio.Lock()
+
+    async def send(self, coro):
+        async with self._lock:
+            result = await coro
+            await asyncio.sleep(self._delay)
+            return result
+
 class PokerBotViewer:
     def __init__(self, bot: Bot):
         self._bot = bot
         self._desk_generator = DeskImageGenerator()
+        self._rate_limiter = RateLimitedSender()
 
     async def send_message_return_id(
         self,
@@ -40,13 +55,15 @@ class PokerBotViewer:
     ) -> Optional[MessageId]:
         """Sends a message and returns its ID, or None if not applicable."""
         try:
-            message = await self._bot.send_message(
-                chat_id=chat_id,
-                parse_mode=ParseMode.MARKDOWN,
-                text=text,
-                reply_markup=reply_markup,
-                disable_notification=True,
-                disable_web_page_preview=True,
+            message = await self._rate_limiter.send(
+                self._bot.send_message(
+                    chat_id=chat_id,
+                    parse_mode=ParseMode.MARKDOWN,
+                    text=text,
+                    reply_markup=reply_markup,
+                    disable_notification=True,
+                    disable_web_page_preview=True,
+                )
             )
             if isinstance(message, Message):
                 return message.message_id
@@ -64,13 +81,15 @@ class PokerBotViewer:
     ) -> Optional[MessageId]:
         for _ in range(2):
             try:
-                message = await self._bot.send_message(
-                    chat_id=chat_id,
-                    parse_mode=parse_mode,  # <--- از پارامتر ورودی استفاده شد
-                    text=text,
-                    reply_markup=reply_markup,
-                    disable_notification=True,
-                    disable_web_page_preview=True,
+                message = await self._rate_limiter.send(
+                    self._bot.send_message(
+                        chat_id=chat_id,
+                        parse_mode=parse_mode,  # <--- از پارامتر ورودی استفاده شد
+                        text=text,
+                        reply_markup=reply_markup,
+                        disable_notification=True,
+                        disable_web_page_preview=True,
+                    )
                 )
                 if isinstance(message, Message):
                     return message.message_id
@@ -84,11 +103,13 @@ class PokerBotViewer:
 
     async def send_photo(self, chat_id: ChatId) -> None:
         try:
-            await self._bot.send_photo(
-                chat_id=chat_id,
-                photo=open("./assets/poker_hand.jpg", 'rb'),
-                parse_mode=ParseMode.MARKDOWN,
-                disable_notification=True,
+            await self._rate_limiter.send(
+                self._bot.send_photo(
+                    chat_id=chat_id,
+                    photo=open("./assets/poker_hand.jpg", 'rb'),
+                    parse_mode=ParseMode.MARKDOWN,
+                    disable_notification=True,
+                )
             )
         except Exception as e:
             print(f"Error sending photo: {e}")
@@ -97,11 +118,13 @@ class PokerBotViewer:
         self, chat_id: ChatId, message_id: MessageId, emoji='🎲'
     ) -> Optional[Message]:
         try:
-            return await self._bot.send_dice(
-                reply_to_message_id=message_id,
-                chat_id=chat_id,
-                disable_notification=True,
-                emoji=emoji,
+            return await self._rate_limiter.send(
+                self._bot.send_dice(
+                    reply_to_message_id=message_id,
+                    chat_id=chat_id,
+                    disable_notification=True,
+                    emoji=emoji,
+                )
             )
         except Exception as e:
             print(f"Error sending dice reply: {e}")
@@ -111,12 +134,14 @@ class PokerBotViewer:
         self, chat_id: ChatId, message_id: MessageId, text: str
     ) -> None:
         try:
-            await self._bot.send_message(
-                reply_to_message_id=message_id,
-                chat_id=chat_id,
-                parse_mode=ParseMode.MARKDOWN,
-                text=text,
-                disable_notification=True,
+            await self._rate_limiter.send(
+                self._bot.send_message(
+                    reply_to_message_id=message_id,
+                    chat_id=chat_id,
+                    parse_mode=ParseMode.MARKDOWN,
+                    text=text,
+                    disable_notification=True,
+                )
             )
         except Exception as e:
             print(f"Error sending message reply: {e}")
@@ -135,15 +160,17 @@ class PokerBotViewer:
             bio.name = 'desk.png'
             im_cards.save(bio, 'PNG')
             bio.seek(0)
-            messages = await self._bot.send_media_group(
-                chat_id=chat_id,
-                media=[
-                    InputMediaPhoto(
-                        media=bio,
-                        caption=caption,
-                    ),
-                ],
-                disable_notification=disable_notification,
+            messages = await self._rate_limiter.send(
+                self._bot.send_media_group(
+                    chat_id=chat_id,
+                    media=[
+                        InputMediaPhoto(
+                            media=bio,
+                            caption=caption,
+                        ),
+                    ],
+                    disable_notification=disable_notification,
+                )
             )
             if messages and isinstance(messages, list) and len(messages) > 0:
                 return messages[0]
@@ -191,13 +218,15 @@ class PokerBotViewer:
     ) -> Optional[MessageId]:
         markup = self._get_cards_markup(cards)
         try:
-            message = await self._bot.send_message(
-                chat_id=chat_id,
-                text="کارت‌های شما " + mention_markdown,
-                reply_markup=markup,
-                reply_to_message_id=ready_message_id,
-                parse_mode=ParseMode.MARKDOWN,
-                disable_notification=True,
+            message = await self._rate_limiter.send(
+                self._bot.send_message(
+                    chat_id=chat_id,
+                    text="کارت‌های شما " + mention_markdown,
+                    reply_markup=markup,
+                    reply_to_message_id=ready_message_id,
+                    parse_mode=ParseMode.MARKDOWN,
+                    disable_notification=True,
+                )
             )
             if isinstance(message, Message):
                 return message.message_id
@@ -248,12 +277,14 @@ class PokerBotViewer:
         markup = self._get_turns_markup(call_check_text, call_check_action)
 
         try:
-            message = await self._bot.send_message(
-                chat_id=chat_id,
-                text=text,
-                reply_markup=markup,
-                parse_mode=ParseMode.MARKDOWN,
-                disable_notification=False,  # player gets notification
+            message = await self._rate_limiter.send(
+                self._bot.send_message(
+                    chat_id=chat_id,
+                    text=text,
+                    reply_markup=markup,
+                    parse_mode=ParseMode.MARKDOWN,
+                    disable_notification=False,  # player gets notification
+                )
             )
             if isinstance(message, Message):
                 return message.message_id
@@ -280,7 +311,9 @@ class PokerBotViewer:
         if not message_id:
             return
         try:
-            await self._bot.edit_message_reply_markup(chat_id=chat_id, message_id=message_id)
+            await self._rate_limiter.send(
+                self._bot.edit_message_reply_markup(chat_id=chat_id, message_id=message_id)
+            )
         except BadRequest as e:
             err = str(e).lower()
             if "message to edit not found" in err or "message is not modified" in err:
@@ -297,7 +330,9 @@ class PokerBotViewer:
         if not message_id:
             return
         try:
-            await self._bot.delete_message(chat_id=chat_id, message_id=message_id)
+            await self._rate_limiter.send(
+                self._bot.delete_message(chat_id=chat_id, message_id=message_id)
+            )
         except BadRequest as e:
             err = str(e).lower()
             if "message to delete not found" in err or "message can't be deleted" in err:
@@ -316,7 +351,9 @@ class PokerBotViewer:
 
         async def _remove():
             try:
-                await self._bot.delete_message(chat_id=chat_id, message_id=message_id)
+                await self._rate_limiter.send(
+                    self._bot.delete_message(chat_id=chat_id, message_id=message_id)
+                )
             except Exception as e:
                 print(f"Could not delete message {message_id} in chat {chat_id}: {e}")
 
@@ -400,12 +437,14 @@ class PokerBotViewer:
         )
         for _ in range(2):
             try:
-                await self._bot.send_message(
-                    chat_id=chat_id,
-                    text=message,
-                    parse_mode=ParseMode.MARKDOWN,
-                    disable_notification=True,
-                    disable_web_page_preview=True,
+                await self._rate_limiter.send(
+                    self._bot.send_message(
+                        chat_id=chat_id,
+                        text=message,
+                        parse_mode=ParseMode.MARKDOWN,
+                        disable_notification=True,
+                        disable_web_page_preview=True,
+                    )
                 )
                 break
             except RetryAfter as e:
